Add PrebalanceTx + TxConstructor now works only with DecoratedTxOuts
We did not used ChainIndexTx in our transaction construction DSL. So, for greater efficiency we removed them. We do not need to query them on the server side.

If there is a need for ChainIndexTx in the future (e.g., to get metadata), we will query and store those separately from DecoratedTxOuts.